+++ src/frame/request_channel.rs
@@ -4,7 +4,7 @@ use super::{Body, Frame, PayloadSupport, Writeable, FLAG_METADATA, REQUEST_MAX};
 use crate::result::RSocketResult;
 use bytes::{BigEndian, BufMut, ByteOrder, Bytes, BytesMut};
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub struct RequestChannel {
   initial_request_n: u32,
   metadata: Option<Bytes>,
@@ -89,7 +89,7 @@ impl Writeable for RequestChannel {
     PayloadSupport::write(bf, self.get_metadata(), self.get_data());
   }
 
-  fn len(&self) -> u32 {
+  fn len(&self) -> usize {
     4 + PayloadSupport::len(self.get_metadata(), self.get_data())
   }
 }
